fix(composer): remove false 20-light caps and roll REST delivery fairly
The "20-light limit" was one literal copied to three places, none of them
Entertainment. Its root was `LightFrame.channelID: UInt8` — a HueStream wire
detail leaked into the shared pure renderer. Because every non-DTLS caller had
to synthesize UInt8 indices, each clamped with `min(…, 20)` to keep the
trapping initialiser safe, and lights 21+ were then dropped silently: not
counted in attemptedOperations, not logged, not surfaced. A 30-light room had
ten bulbs that never received a single frame for the life of the composition
while the tray reported a healthy cadence.

The genuine DTLS path has no 20 anywhere. `validatedChannelIDs` already
validates the whole ordered channel set all-or-nothing, and the only honest
light limit in the product is the bridge's own 10-per-entertainment-area.

- `LightFrame.channelID` and `render`/`renderMixed` take Int render indices.
  Wire ids are never reconstructed from render output; the DTLS send zips the
  already-validated `[UInt8]` positionally and sends nothing at all rather
  than a partial or misaligned frame.
- `GradientChannelMap.channelBudget` deleted. Every light keeps an entry, and
  a strip expands to its own `points_capable` — it could previously be starved
  to 1 channel by unrelated bulbs, silently losing the gradient.
- REST rooms larger than one sweep now rotate instead of truncating.
  Partitions are contiguous and NON-WRAPPING, so each rotation is an exact
  ordered partition: 21 lights go 20 then 1, 64 go 20/20/20/4, and every
  eligible operation is served exactly once per rotation. The cursor advances
  from `outcomes.count` after each task group — the same evidence Packet 4's
  telemetry uses, so the two can never disagree about what was dispatched.
- Rotation and degradation state are keyed by the exact (bridgeKey, scope)
  identity Packet 4 established, never roomID alone, and share that session's
  lifecycle. The roomID-keyed `compositionRuntimes` limitation is pre-existing
  and is NOT solved here.
- Quiescence requires a rotation that DELIVERED, not one that merely
  attempted. A failure anywhere leaves the room re-rotating, so completion-only
  bookkeeping can no longer be overruled by a gate told the rotation was
  "complete". A persistently failing light keeps a static room rotating — that
  is stated, not hidden; backoff is later work.
- Gradient metadata is re-indexed into render-channel order. The spatial,
  radial and angular arrays were per-physical-light while `render` indexed
  them per render channel, so every light after a strip read a neighbour's
  geometry — masked only by the small totals the old budget enforced.
- Transport copy: `roomModeStatus` is pure and total over (fallback reason ×
  large room), so a capacity refusal into a big room states both facts instead
  of one erasing the other. No sentence claims a light count.
- Two user-facing "REST" leaks fixed. One was known; the other — "🔌 REST" in
  the Perform header — was found by the source-inspection guard added here,
  which both existing guards had missed.

Guard 8 fails the build if a `UInt8(min(` clamp or `channelBudget` returns to
the render or bridge-stored paths; verified against a deliberate regression.

Full suite: 1033 passed, 0 failed. hardening_guards: all passed.

+++ HueHomeTests/BeatMathTests.swift
@@ -183,7 +183,7 @@ final class BeatMathTests: XCTestCase {
 final class CompositionMixerTests: XCTestCase {
 
     private let snap120 = BeatSnapshot(bpm: 120, beatEpoch: 0, beatsPerBar: 4)
-    private let channels: [UInt8] = [0, 1, 2]
+    private let channels: [Int] = [0, 1, 2]
 
     private func box(hueShift: Double = 0) -> CompositionParamBox {
         let box = CompositionParamBox(
